% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-badge.R
\name{material_badge}
\alias{material_badge}
\alias{update_material_badge}
\title{Create a materialize badge}
\usage{
material_badge(inputId, label = NULL, color = NULL)

update_material_badge(inputId, label = NULL, class = "new",
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The badge label}

\item{color}{String. The color name of the badge Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize badge
}
\examples{
require(materializer)
material_badge(
  inputId = "example_button",
  label = "4",
  color = "blue lighten-2"
)
}
