% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-collapsible.R
\name{material_collapsible}
\alias{material_collapsible}
\alias{update_material_collapsible}
\title{The material Collapsible UI Element}
\usage{
material_collapsible(inputId, cllist, class = "accordion",
  headercolor = "white", bodycolor = "white", color = NULL)

update_material_collapsible(inputId, cllist,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{cllist}{List. The collapsible element list. The list need to contain a 'name' and 'content' element, in adition could contain 'icon', 'badge' and 'active' (logical) elements.}

\item{class}{String. A class type for UI Collapsible. Values should be 'accordion', 'expandable' or 'popout'.}

\item{headercolor}{String. The color name of header. Leave empty for 'white' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{bodycolor}{String. The color name of body. Leave empty for 'white' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{color}{String. The color name of the text and icons. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
The material Collapsible UI Element
}
\examples{
require(materializer)
material_collapsible(
inputId = "example_collapsible_menu",
  cllist = list(
    c(target = "tgt_1", name = "One"),
    c(target = "tgt_2", name = "Two"),
    c(target = "tgt_3", name = "Three", icon = "cloud"),
    c(target = "tgt_4", icon = "alarm")
 )
)
}
