% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-colormap.R
\name{material_colormap}
\alias{material_colormap}
\title{Convert color material to hexcode color}
\usage{
material_colormap(color)
}
\arguments{
\item{color}{String. The input identifier used to access the value.}
}
\description{
Color material to hexcode color
}
