% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-grid.R
\name{material_container}
\alias{material_container}
\title{Create a container to organize UI content}
\usage{
material_container(...)
}
\arguments{
\item{...}{tagList. UI content list}
}
\description{
UI content can be insert in a container to organize items on a page.
}
\examples{
material_container()
}
