% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-date-picker.R
\name{material_date_picker}
\alias{material_date_picker}
\title{Create a materialize date picker}
\usage{
material_date_picker(inputId, label, color = NULL)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The date picker label.}

\item{color}{String. The hex codes color of the picker. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{This input requires using color hex codes, rather than the word form. E.g., "#ef5350", rather than "red lighten-1".}}
}
\description{
Build a materialize date picker.
}
\examples{
material_date_picker(
  inputId = "example_date_picker",
  label = "Date Picker"
)
}
