% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-grid.R
\name{material_divider}
\alias{material_divider}
\title{Create a divider to organize UI content}
\usage{
material_divider()
}
\description{
UI content can be divide to organize items on a page.
}
\examples{
require(materializer)
material_divider()
}
