% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-dropdown.R
\name{material_dropdown}
\alias{material_dropdown}
\alias{update_material_dropdown}
\title{The material Dropdown Menu UI Element}
\usage{
material_dropdown(inputId, button, ddlist, bgcolor = "white", color = NULL)

update_material_dropdown(inputId, ddlist,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{button}{String. The tooltip message.}

\item{ddlist}{List. The dropdown element list.The list need to contain a 'target' element and in adition could contain 'name' and 'icon' elements.}

\item{bgcolor}{String. The color name for Dropdown backgroud. Leave empty for 'white' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{color}{String. The color name of the Checkbox. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
The material Dropdown Menu UI Element
}
\examples{
require(materializer)
material_dropdown(
inputId = "example_dropdown_menu",
  button = material_button(
    inputId = "example_button",
    label = "Drop me",
    icon = "arrow_drop_down"
  ),
  ddlist = list(
    c(target = "tgt_1", name = "One"),
    c(target = "tgt_2", name = "Two"),
    NULL,
    c(target = "tgt_3", name = "Three", icon = "cloud"),
    c(target = "tgt_4", icon = "alarm")
 )
)
}
