% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-fab.R
\name{material_fab}
\alias{material_fab}
\title{Create a materialize fab action button (fab)}
\usage{
material_fab(inputId, actlist = list(), icon = "mode_edit",
  size = "large", color = NULL, icolor = NULL, direction = "top",
  position = "bottom-right", fixed = TRUE, hover = TRUE)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{actlist}{List. A list of action attributes such as target, icon, size, color and icolor.}

\item{icon}{String The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{size}{String. The size of the icon. Values should be 'large', 'medium' or 'small'.}

\item{color}{String. The color name of the fab action button. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{icolor}{String. The color name of the icon. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{direction}{String. The direction of the fab action button when open. values should be 'top', 'right', 'buttom', 'left'.}

\item{position}{String. The position of the fab action button. Values should be 'bottom-left', 'bottom-right', 'top-left' and 'top-right'.}

\item{fixed}{Logical. The fab action button should be fixed?}

\item{hover}{Logical. Should the fab action button be hoveable ?}
}
\description{
Build a materialize button. The initial value is zero, and increases by one on each press.
}
\examples{
require(materializer)
material_fab(
  inputId = "example_flt_action",
  actlist = list(
    c(target = "ex1", icon = "home"),
    c(target = "ex2", icon = "chart")
  ),
  icon = "add",
  color = 'teal lighten-1'
)
}
