% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-floating.R
\name{material_floating}
\alias{material_floating}
\title{Create a materialize floating action button}
\usage{
material_floating(inputId, icon, size = "large", color = NULL,
  direction = "left", action = list(target = NULL, icon = NULL, size = NULL,
  color = NULL))
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{icon}{String list. The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{size}{String. The size of the icon. Values should be 'large', 'medium' or 'small'.}

\item{color}{String. The hex codes color of the check. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{direction}{String. The direction of the floating action button when open. values should be 'top', 'right', 'buttom', 'left'.}

\item{action}{List. A list of action attributes such as target, icon, size and color.}
}
\description{
Build a materialize button. The initial value is zero, and increases by one on each press.
}
\examples{
material_floating(
  inputId = "example_flt_action",
  icon = "add",
  color = "#ffd600",
  action = list(
    target = c("id1", "id2", "id3", "id4", "id5"),
    icon = c("model_edit", "insert_chart", "attach_file", "publish", "format_quote")
  )
)
}
