% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize_image.R
\name{material_img}
\alias{material_img}
\title{Material Image}
\usage{
material_img(src, width = NULL, heigth = NULL, class = NULL,
  caption = NULL, ...)
}
\arguments{
\item{src}{String. The file path of the image}

\item{width}{Integer The width of the image in pixels}

\item{heigth}{Integer The heigth of the image in pixels}

\item{class}{String. Aditional class to image}

\item{caption}{String. The caption of the image}

\item{...}{tagList. Any additional tag list}
}
\description{
Createa material Image element
}
\examples{
require(materializer)
material_img(src = "image.png", width = 50)
}
