% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-like.R
\name{material_like}
\alias{material_like}
\alias{update_material_like}
\title{Create a materialize like element}
\usage{
material_like(inputId, liked = NULL, size = "medium",
  disabled = FALSE)

update_material_like(inputId, liked = NULL, disabled = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{liked}{Boolian. The initial status of the element. Could be TRUE, FALSE or NULL}

\item{size}{String. Size of icon, could be 'tiny' (1rem), 'small' (2rem), 'medium' (4rem) and 'large' (6rem).}

\item{disabled}{Boolean. Disable iteractivity (click).}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize like and dislike element
}
\examples{
require(materializer)
material_like(
  inputId = "example_checkbox",
  liked = TRUE
)
}
