% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-meter.R
\name{material_meter}
\alias{material_meter}
\alias{update_material_meter}
\title{Create a materialize meter}
\usage{
material_meter(inputId, value = 0.5, min = 0, max = 1, high = 0.8,
  low = 0.2, optimum = 0.5, width = "5em", height = "1em", ...)

update_material_meter(inputId, value = NULL, min = NULL, max = NULL,
  high = NULL, low = NULL, optimum = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{value}{Double. A floating point number that represents the current value of the measured range. This must be between the min and the max value (if specified).}

\item{min}{Double Indicates the lower bound of the measured range. This must be less than the max value (if specified). If unspecified, the minimum value is 0.}

\item{max}{Double. Indicates the upper bound of the measured range. This must be greater than the min value (if specified). If unspecified, the maximum value is 1.0.}

\item{high}{Double. It represents the lower bound of the high end of the measured range. This must be less than the max attribute, but greater than the low and min value (if specified). If unspecified, or if greater than the max value, the high value is equal to the max value.}

\item{low}{Double. It represents the upper bound of the low end of the measured range. This must be greater than the min attribute, but less than the high and max value (if specified). If unspecified, or if less than the minimum value, the low value is equal to the min value.}

\item{optimum}{Double. This attribute indicates the optimum value and must be within the range of min and max values. When used with the low and high attribute, it indicates the preferred zone for a given range.}

\item{width}{String. The width of the meter. Could be any css length.}

\item{height}{String. Could be any css length}

\item{...}{tagList. Aditional arguments}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize rating
}
\examples{
require(materializer)
material_meter(
  inputId = "examplemeter",
  value = 0.5,
  min = 0,
  max = 1
)
}
