% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-modal.R
\name{material_modal}
\alias{material_modal}
\title{Place UI content in a modal}
\usage{
material_modal(inputId, ..., btnShow = material_button(paste0(inputId,
  "_show"), "Show"), btnClose = material_button(paste0(inputId, "_close"),
  "Close", class = "flat"), fixedFooter = FALSE)
}
\arguments{
\item{inputId}{String. The ID for the modal. Must be unique per application.}

\item{...}{The UI elements to place in the modal}

\item{btnShow}{Button. The modal show button tag.}

\item{btnClose}{Button. The modal close button tag.}

\item{fixedFooter}{Logical. Should the footerbe fixed?}
}
\description{
Put any UI object inside of a modal. The modal will open when the button is pressed.
}
\examples{
material_modal(
  inputId = "example_modal",
  shiny::tags$p("Modal Content")
)
}
