% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-radio.R
\name{material_radio}
\alias{material_radio}
\alias{update_material_radio}
\title{Create a materialize radio button}
\usage{
material_radio(inputId, choices, selected = NULL, disabled = NULL,
  class = "with-gap", inline = FALSE, color = NULL)

update_material_radio(inputId, choices = NULL, selected = NULL,
  disabled = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{choices}{Named vector. The radio names and underyling values.}

\item{selected}{String. The initially selected underyling value.}

\item{disabled}{String. The radio button should be disabled?}

\item{class}{String. Aditional class for checkbox. Values should be with-gap or NULL}

\item{inline}{Logical. The option to create the radio buttons inline.}

\item{color}{String. The color name of the radio buttons. Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize radio button.
}
\examples{
require(materializer)
material_radio(
  inputId = "example_radio_button",
  choices = c(
    "Cake" = "c",
    "Pie" = "p",
    "Brownie" = "b"
  ),
  color = 'teal lighten-1'
)
}
