% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-rating.R
\name{material_rating}
\alias{material_rating}
\alias{update_material_rating}
\title{Create a materialize rating}
\usage{
material_rating(inputId, n = 5L, checked = 3, size = 1.25,
  space = 3, disabled = FALSE, color = NULL)

update_material_rating(inputId, checked = NULL, disabled = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{n}{integer The number of stars}

\item{checked}{Double The rating value, cloud be half values.}

\item{size}{Double. The size of the star rating based on 'em' css}

\item{space}{Double. THe space between the stars based on pixels.}

\item{disabled}{Boolean. Is the rating disable?}

\item{color}{String. The color name of the Checkbox. Leave empty for the 'yellow accent-4' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize rating
}
\examples{
require(materializer)
material_rating(
  inputId = "examplerating",
  n = 5,
  checked = 3.5,
  color = 'teal lighten-1'
)
}
