% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-column.R
\name{material_section}
\alias{material_section}
\title{Create a section to organize UI content}
\usage{
material_section(..., title = NULL)
}
\arguments{
\item{...}{The UI elements to place in the column.}

\item{title}{Integer. The width of the column. The value should be between 1 and 12.}
}
\description{
UI content can be insert in section to organize items on a page.
}
\examples{
material_section(
  title = "Section",
  shiny::h1("Content goes here")
)
}
