% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-select.R
\name{material_select}
\alias{material_select}
\alias{update_material_select}
\title{Create a materialize select}
\usage{
material_select(inputId, label, choices = NULL, selected = NULL,
  multiple = FALSE, color = NULL)

update_material_select(inputId, choices = NULL, selected = NULL,
  label = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The dropdown label.}

\item{choices}{Named vector. The option names and underyling values.}

\item{selected}{String. The initially selected underyling value.}

\item{multiple}{Boolean. Can multiple items be selected?}

\item{color}{String. The color name of the select input. Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain}
}
\description{
Build a materialize select
}
\examples{
material_select(
  inputId = "example_dropdown",
  label = "Drop down",
  choices = c(
    "Chicken" = "c",
    "Steak" = "s",
    "Fish" = "f"
  ),
  selected = c("c"),
  multiple = FALSE,
  color = 'teal lighten-1'
)
}
