% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-sidenav.R
\name{material_sidenav}
\alias{material_sidenav}
\title{The material SideNav bar}
\usage{
material_sidenav(inputId, navlist, logo = NULL, fixed = TRUE, width = 300,
  bgcolor = "white", color = NULL)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{navlist}{list. The nav bar element list. The list need to contain a 'target' element and in adition could contain 'name' (string), 'icon' (string), 'active' (logical).}

\item{logo}{list. The nav bar logo element list. The list need to contain a 'target' element and in adition could contain 'name' (string), 'icon' (string) or any tagList elements.}

\item{fixed}{Logical Should the sidenav be fixed?}

\item{width}{Integer. The width size of the sizebar}

\item{bgcolor}{String. The color name for Sidenav backgroud. Leave empty for 'white' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{color}{String. The color name for links and icons in Sidenav Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}
}
\description{
The material SideNav bar UI Element
}
\examples{
require(materializer)
material_sidenav(
  inputId = "example_navbar",
  navlist = list(
    c(target = "lisk_1.html", name = "One"),
    c(target = "lisk_2.html", name = "Two"),
    c(target = "lisk_3.html", name = "Three", icon = "cloud"),
    c(target = "lisk_4.html", icon = "alarm", active = TRUE)
  )
)
}
