% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-slider.R
\name{material_slider}
\alias{material_slider}
\title{Create a materialize slider}
\usage{
material_slider(inputId, label, min_value = 0, max_value = 100,
  initial_value = min_value, color = NULL)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The slider label.}

\item{min_value}{Number. The minimum value on the slider.}

\item{max_value}{Number. The maximum value on the slider.}

\item{initial_value}{Number. The initial value of the slider.}

\item{color}{String. The slider color hex codes. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{This input requires using color hex codes, rather than the word form. E.g., "#ef5350", rather than "red lighten-1".}}
}
\description{
Build a materialize slider.
}
\examples{
material_slider(
  inputId = "example_slider",
  label = "slider",
  min_value = 5,
  max_value = 15,
  initial_value = 10,
  color = "#ef5350"
)
}
