% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-tabs.R
\name{material_tab_content}
\alias{material_tab_content}
\title{Place UI content within a tab content}
\usage{
material_tab_content(inputId, ...)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{...}{tagList. The UI elements to be insertin the tab content}
}
\description{
Use this function to create a tab content in your application.
}
