% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-tags.R
\name{material_tags}
\alias{material_tags}
\alias{update_material_tags}
\title{Material Tags (Chips)}
\usage{
material_tags(inputId, tags = list(), placeholder = NULL,
  autocomplete = NULL, color = NULL, ...)

update_material_tags(inputId, addtags = NULL, deltags = NULL,
  seltags = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String.}

\item{tags}{list A list of tag elements, could have 'tag', 'icon', 'image'.}

\item{placeholder}{String. A placeholder for the input text.}

\item{autocomplete}{String. A list of words to autocomplete}

\item{color}{String. The color name of the text input. Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{...}{tagList. Any additional tag list}

\item{addtags}{list. Tags to be added, list of tags with elements 'tag'.}

\item{deltags}{list. Tags to be deleted, list of positions}

\item{seltags}{list. Tags to be selected, list of positions}

\item{session}{Shiny default reactive domain.}
}
\description{
Createa material tag element
}
\examples{
require(materializer)
material_tags(
  inputId = "extags",
  tags = list(c(tag = "Apple", icon = "close")),
  placeholder = "BigTech Companies"
)
}
