% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-text.R
\name{material_text}
\alias{material_text}
\alias{update_material_text}
\title{Create a materialize text box}
\usage{
material_text(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, type = "text", icon = NULL,
  color = NULL)

update_material_text(inputId, label = NULL, placeholder = NULL,
  readonly = NULL, help = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The text input label.}

\item{placeholder}{String. The placeholder of the text input}

\item{inline}{Logical. The text input should be inline with others elements?. Value should be TRUE or FALSE}

\item{readonly}{Logical. The text input should be read-only?. Value should be TRUE or FALSE}

\item{help}{String. The helper text below text input.}

\item{type}{String. The type of the text input. Valiue should be text, email, password.}

\item{icon}{String. The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{color}{String. The hex code color of the text box. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{This input requires using color hex codes, rather than the word form. E.g., "#ef5350", rather than "red lighten-1".}}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize text input
}
\examples{
material_text(
  inputId = "example_text",
  label = "Label",
  color = "#ef5350"
)
}
