% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-tooltip.R
\name{material_tooltip}
\alias{material_tooltip}
\title{Include a tooltip in any tga list}
\usage{
material_tooltip(tagList, tooltip, position = "bottom", delay = 50)
}
\arguments{
\item{tagList}{tagList The UI elements}

\item{tooltip}{String. The tooltip message.}

\item{position}{String. The position of the tooltip. Values should be 'top', 'right', 'bottom', 'left'.}

\item{delay}{Integer. The delay time before tooltip desappears.}
}
\description{
Include a tooltip in any taglist
}
\examples{
require(materializer)
material_tooltip(
  material_button(
    inputId = "example_button",
    label = "Button",
    icon = "cloud"
  ),
  tooltip = "Open the Cloud"
)
}
