% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/materialize.R
\name{materialize}
\alias{materialize}
\title{Add \href{http://materializecss.com/}{materialize} to your R}
\usage{
materialize(..., width = 0, height = 0)
}
\arguments{
\item{...}{expansion room.  The \code{...} currently are a placeholder for future functionality.}

\item{width}{\code{integer} width in pixels for container.  \code{width = 0} since the expectation is
for the container to be an invisible byproduct.}

\item{height}{\code{integer} height in pixels for container.  \code{height = 0} since the expectation is
         for the container to be an invisible byproduct.}
}
\description{
One and only function offered by \code{materializeR}.  The \code{...} is offered but really not all that helpful
currently except for one secret power.  Use of the package \code{htmltools}
from RStudio is a highly encouraged ally for \code{materializeR}.
}

