% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-breadcrumb.R
\name{material_breadcrumb}
\alias{material_breadcrumb}
\title{Create a breadcrumb}
\usage{
material_breadcrumb(inputId, navlist)
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{navlist}{navlist. The navigation list, should contain target and name elements and optional as icon.}
}
\description{
UI content can be insert in a container to organize items on a page.
}
\examples{
require(materializer)
material_breadcrumb(inputId = "breadcrumb", navlist = list(page1 = list(target="!#", name="Page 1")))
}
