% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-card.R
\name{material_card}
\alias{material_card}
\title{Create a card that will contain UI content}
\usage{
material_card(..., title = NULL, img = NULL, footer = NULL,
  class = "stacked", header = TRUE, depth = NULL, horizontal = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the card}

\item{title}{String. The title of the card}

\item{img}{String. The filepath of a image of the card}

\item{footer}{The UI elements to place in the card footer}

\item{class}{The aditional class to the card. Values shound be stacked or panel.}

\item{header}{Option to include a header in the card.}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{horizontal}{Should the card be on horizontal orientation?}
}
\description{
UI content can be placed in cards to organize items on a page.
}
\examples{
material_card(
  title = "Example Card",
  shiny::h5("Card Content"),
  footer = shiny::hr()
)
}
