% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-card.R
\name{material_card}
\alias{material_card}
\alias{update_material_card}
\title{Create a card that will contain UI content}
\usage{
material_card(inputId, content = NULL, title = NULL, img = NULL,
  footer = NULL, class = "stacked", header = TRUE, tabs = NULL,
  depth = NULL, horizontal = FALSE, hoverable = TRUE,
  bgcolor = NULL, ...)

update_material_card(inputId, content = NULL, title = NULL,
  footer = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{content}{tagList. The UI elements to place in card content}

\item{title}{String. The title of the card}

\item{img}{String. The filepath of a image of the card}

\item{footer}{tagList. The UI elements to place in the card footer}

\item{class}{String. The aditional class to the card. Values should be stacked or panel.}

\item{header}{Logical. Option to include a header in the card.}

\item{tabs}{tgList. A material_tabs object to insert inside the card}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{horizontal}{Logical. Should the card be on horizontal orientation?'}

\item{hoverable}{Loginal. Enable/disable hover effect.}

\item{bgcolor}{String. The background color name of the card. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{...}{tagList. additional tag elements}

\item{session}{Shiny default reactive domain.}
}
\description{
UI content can be placed in cards to organize items on a page.
}
\examples{
require(materializer)
material_card(
  inputId = "exampleCard",
  title = "Example Card",
  content = shiny::h5("Card Content"),
  footer = shiny::hr()
)
}
