% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-checkbox.R
\name{material_checkbox}
\alias{material_checkbox}
\alias{update_material_checkbox}
\title{Create a materialize checkbox}
\usage{
material_checkbox(inputId, label, checked = FALSE, class = "filled-in",
  color = NULL)

update_material_checkbox(inputId, label = NULL, checked = FALSE,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The checkbox label.}

\item{checked}{Boolean. Is the checkbox initially checked?}

\item{class}{String. Aditional class for checkbox. Values should be filled-in or NULL}

\item{color}{String. The color name of the Checkbox. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize checkbox. The value is a boolean (TRUE if checked, FALSE if not checked).
}
\examples{
material_checkbox(
  inputId = "example_checkbox",
  label = "Checkbox",
  checked = TRUE,
  color = 'teal lighten-1'
)
}
