% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-like.R
\name{material_like}
\alias{material_like}
\alias{update_material_like}
\title{Create a materialize like element}
\usage{
material_like(inputId, liked = TRUE, size = "medium", color = NULL)

update_material_like(inputId, liked = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{liked}{Boolian. The initial status of the element}

\item{size}{String. Size of icon, could be 'tiny' (1rem), 'small' (2rem), 'medium' (4rem) and 'large' (6rem).}

\item{color}{String. The color name of the Checkbox. Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize like and dislike element
}
\examples{
require(materializer)
material_like(
  inputId = "example_checkbox",
  color = 'teal lighten-1'
)
}
