% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-page.R
\name{material_page}
\alias{material_page}
\title{Create a materialize page}
\usage{
material_page(..., title = "", footer = NULL, bgcolor = "grey lighten-4",
  fontcolor = NULL, localfonts = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the page.}

\item{title}{String. The title of the page.}

\item{footer}{tagList. The footerof thepage}

\item{bgcolor}{Page background color. Leave blank for "grey lighten-4" color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{fontcolor}{String. The title font color. Leave blank for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{localfonts}{Boolean. Should the material font files be included? (This will place the font sources in a directory 'www', at the same location as the app code.)}
}
\description{
Build a materialize page.
}
\examples{
material_page(
  title = "Example Title",
  bgcolor = "blue lighten-4",
  shiny::tags$h1("Page Content")
)
}
