% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-slider.R
\name{material_slider}
\alias{material_slider}
\alias{update_material_slider}
\title{Create a materialize slider}
\usage{
material_slider(inputId, label, minval = 0, maxval = 100,
  initval = minval, color = NULL)

update_material_slider(inputId, label = NULL, minval = NULL,
  maxval = NULL, initval = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The slider label.}

\item{minval}{Number. The minimum value on the slider.}

\item{maxval}{Number. The maximum value on the slider.}

\item{initval}{Number. The initial value of the slider.}

\item{color}{String. The color name of the slider Leave empty for the 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize slider.
}
\examples{
material_slider(
  inputId = "example_slider",
  label = "slider",
  minval = 5,
  maxval = 15,
  initval = 10,
  color = 'teal lighten-1'
)
}
