% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-switch.R
\name{material_switch}
\alias{material_switch}
\alias{update_material_switch}
\title{Create a materialize switch}
\usage{
material_switch(inputId, on = "On", off = "Off", value = FALSE,
  disabled = FALSE, color = NULL)

update_material_switch(inputId, on = NULL, off = NULL, value = NULL,
  disabled = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{on}{String. The label for the 'on' position of the switch.}

\item{off}{String. The label for the 'off' position of the switch.}

\item{value}{Boolean. Is the switch initially on?}

\item{disabled}{String. The radio button should be disabled?}

\item{color}{String. The color name of the Switch. Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize switch. The value is a boolean (TRUE if 'on', FALSE if 'off').
}
\examples{
require(materializer)
material_switch(
  inputId = "example_switch",
  off = "Off",
  on = "On",
  value = TRUE,
  color = 'teal lighten-1'
)
}
