% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize-text.R
\name{material_text}
\alias{material_text}
\alias{material_number}
\alias{material_email}
\alias{material_password}
\alias{material_date}
\alias{material_time}
\alias{update_material_text}
\alias{update_material_number}
\alias{update_material_email}
\alias{update_material_password}
\alias{update_material_date}
\title{Create a materialize text box}
\usage{
material_text(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, type = "text", class = NULL,
  icon = NULL, color = NULL, error = "wrong", success = "valid")

material_number(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, class = NULL, icon = NULL,
  color = NULL)

material_email(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, class = NULL, icon = NULL,
  color = NULL)

material_password(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, class = NULL, icon = NULL,
  color = NULL)

material_date(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, class = NULL, icon = NULL,
  color = NULL)

material_time(inputId, label, placeholder = NULL, inline = FALSE,
  readonly = FALSE, help = NULL, class = NULL, icon = NULL,
  color = NULL)

update_material_text(inputId, value = NULL, label = NULL,
  placeholder = NULL, readonly = NULL, help = NULL,
  session = shiny::getDefaultReactiveDomain())

update_material_number(inputId, value = NULL, label = NULL,
  placeholder = NULL, readonly = FALSE, help = NULL,
  session = shiny::getDefaultReactiveDomain())

update_material_email(inputId, value = NULL, label = NULL,
  placeholder = NULL, readonly = FALSE, help = NULL,
  session = shiny::getDefaultReactiveDomain())

update_material_password(inputId, value = NULL, label = NULL,
  placeholder = NULL, readonly = FALSE, help = NULL,
  session = shiny::getDefaultReactiveDomain())

update_material_date(inputId, value = NULL, label = NULL,
  placeholder = NULL, readonly = FALSE, help = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{label}{String. The text input label.}

\item{placeholder}{String. The placeholder of the text input}

\item{inline}{Logical. The text input should be inline with others elements?. Value should be TRUE or FALSE}

\item{readonly}{Logical. The text input should be read-only?. Value should be TRUE or FALSE}

\item{help}{String. The helper text below text input.}

\item{type}{String. The type of the text input. Valiue should be text, number, email, password, time, date}

\item{class}{String. Additional class for text input.}

\item{icon}{String. The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{color}{String. The color name of the text input. Leave empty for 'teal lighten-1' color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{error}{String. Validate mensage when is not valid}

\item{success}{String. Validate mensage when is  valid}

\item{value}{String or Numeric. The value to update the input.}

\item{session}{Shiny default reactive domain.}
}
\description{
Build a materialize text input
}
\examples{
require(materializer)
material_text(
  inputId = "example_text",
  label = "Label",
  color = 'teal lighten-1'
)
}
